rename to mavsdk and update proto submodule
