+++ examples/android-client/app/build.gradle
@@ -13,7 +13,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "io.dronecore.dronecoreclient"
+        applicationId "io.mavsdk.androidclient"
         minSdkVersion 21
         targetSdkVersion 28
         versionCode 1
@@ -56,7 +56,7 @@ dependencies {
     implementation 'com.google.android.gms:play-services-maps:16.1.0'
     implementation 'com.google.android.material:material:1.0.0'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.0.0'
-    implementation 'org.dronecode.sdk:dronecode-sdk:0.1.0'
+    implementation 'io.mavsdk:mavsdk:0.1.0'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test:runner:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.1'
