+++ examples/android-client/app/src/test/java/io/mavsdk/androidclient/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package io.dronecore.dronecoreclient;
+package io.mavsdk.androidclient;
 
 import org.junit.Test;
