+++ sdk/build.gradle
@@ -13,7 +13,7 @@ subprojects {
     apply plugin: 'idea'
     apply plugin: 'java-library'
 
-    group = 'org.dronecode.sdk'
+    group = 'io.mavsdk'
     version = '0.1.0'
 
     repositories {
