--- sdk/dronecode-sdk/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'dronecode-sdk'
