+++ sdk/settings.gradle
@@ -1,1 +1,1 @@
-include 'dronecode-sdk'
+include 'mavsdk'
