+++ sdk/templates/file.j2
@@ -1,4 +1,4 @@
-package io.dronecode_sdk.{{ plugin_name.lower_camel_case }};
+package io.mavsdk.{{ plugin_name.lower_camel_case }};
 
 import io.grpc.ManagedChannel;
 import io.grpc.okhttp.OkHttpChannelBuilder;
