+++ sdk/templates/struct.j2
@@ -18,7 +18,11 @@ public static class {{ name.upper_camel_case }} {
 
     {%- for field in fields %}
     {%- if field.type_info.is_repeated %}
-      .addAll{{ field.name.upper_camel_case}}({{ field.name.lower_camel_case }}.stream().map({{ field.type_info.inner_name }}::rpc{{ field.type_info.inner_name }}).collect(Collectors.toList()))
+        {%- if field.type_info.is_primitive %}
+      .addAll{{ field.name.upper_camel_case }}({{ field.name.lower_camel_case }})
+        {%- else %}
+      .addAll{{ field.name.upper_camel_case }}({{ field.name.lower_camel_case }}.stream().map({{ field.type_info.inner_name }}::rpc{{ field.type_info.inner_name }}).collect(Collectors.toList()))
+        {%- endif %}
     {%- else %}
         {%- if field.type_info.is_primitive %}
       .set{{ field.name.upper_camel_case }}(this.{{ field.name.lower_camel_case }})
@@ -33,7 +37,11 @@ public static class {{ name.upper_camel_case }} {
     return new {{ name.upper_camel_case }}(
             {%- for field in fields %}
             {%- if field.type_info.is_repeated %}
+                {%- if field.type_info.is_primitive %}
+                rpc{{ name.upper_camel_case }}.get{{ field.name.upper_camel_case }}List()
+                {%- else %}
                 rpc{{ name.upper_camel_case }}.get{{ field.name.upper_camel_case }}List().stream().map({{ field.type_info.inner_name }}::translateFromRpc).collect(Collectors.toList())
+                {%- endif %}
             {%- else %}
                 {%- if field.type_info.is_primitive %}
             rpc{{ name.upper_camel_case }}.get{{ field.name.upper_camel_case }}(){{ "," if not loop.last }}
